scripts/indexer/core/requirements.txt
pyairtable==2.3.3
langchain==0.3.15
langchain-openai==0.2.8
langchain-anthropic==0.3.15
openai==1.55.3
anthropic>=0.52.0
supabase==2.9.1
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
tiktoken==0.8.0
numpy==1.24.3
pandas==2.1.4
requests==2.31.0
schedule==1.2.0 